codegen sort by msg name, fix example

+++ examples/read_bag/src/main.rs
@@ -1,4 +1,4 @@
-use std::f64::consts::PI;
+use std::f32::consts::PI;
 use std::{fs::File, io::Write, path::PathBuf};
 
 use frost::query::Query;
@@ -54,7 +54,7 @@ fn main() {
     println!("Last {} message is {}", &msg_view.topic, msg.data);
 
     // check msg constants (type not in bag)
-    assert_eq!(dummy_msgs::Dummy::PI, PI as f32);
+    assert_eq!(dummy_msgs::Dummy::PI, PI);
     assert_eq!(dummy_msgs::Dummy::N_PI, -PI);
     assert_eq!(dummy_msgs::Dummy::HELLO, "WORLD");
